+++ src/lib/pages/contract-details/components/InstantiateInfo.tsx
@@ -1,4 +1,4 @@
-import { Flex, Heading, Text } from "@chakra-ui/react";
+import { Divider, Flex, Text } from "@chakra-ui/react";
 
 import { ExplorerLink } from "lib/components/ExplorerLink";
 import { LabelText } from "lib/components/LabelText";
@@ -23,16 +23,43 @@ export const InstantiateInfo = ({ contractData }: InstantiateInfoProps) => {
         <>
           <LabelText label="Network">{contractData.chainId}</LabelText>
 
+          <LabelText
+            label="From Code"
+            helperText1={contractData.codeInfo?.description}
+          >
+            <ExplorerLink
+              type="code_id"
+              value={contractData.instantiateInfo.codeId}
+              canCopyWithHover
+            />
+          </LabelText>
+
+          {contractData.instantiateInfo.admin && (
+            <LabelText label="Admin Address">
+              <ExplorerLink
+                type="user_address"
+                value={contractData.instantiateInfo.admin}
+                canCopyWithHover
+              />
+            </LabelText>
+          )}
+
+          <Divider border="1px solid" borderColor="divider.main" />
+
           {contractData.instantiateInfo &&
             (contractData.instantiateInfo.createdHeight !== -1 ? (
               <LabelText
                 label="Instantiated Block Height"
-                helperText1={formatUTC(
+                helperText1={
                   contractData.instantiateInfo.createdTime
-                )}
-                helperText2={dateFromNow(
+                    ? formatUTC(contractData.instantiateInfo.createdTime)
+                    : undefined
+                }
+                helperText2={
                   contractData.instantiateInfo.createdTime
-                )}
+                    ? dateFromNow(contractData.instantiateInfo.createdTime)
+                    : undefined
+                }
               >
                 <ExplorerLink
                   value={contractData.instantiateInfo.createdHeight.toString()}
@@ -50,17 +77,6 @@ export const InstantiateInfo = ({ contractData }: InstantiateInfoProps) => {
             <ExplorerLink
               type="user_address"
               value={contractData.instantiateInfo.instantiator}
-              canCopyWithHover
-            />
-          </LabelText>
-
-          <LabelText
-            label="From Code"
-            helperText1={contractData.codeInfo?.description}
-          >
-            <ExplorerLink
-              type="code_id"
-              value={contractData.instantiateInfo.codeId}
               canCopyWithHover
             />
           </LabelText>
@@ -86,15 +102,6 @@ export const InstantiateInfo = ({ contractData }: InstantiateInfoProps) => {
             </LabelText>
           )}
 
-          {contractData.instantiateInfo.admin && (
-            <LabelText label="Admin Address">
-              <ExplorerLink
-                type="user_address"
-                value={contractData.instantiateInfo.admin}
-              />
-            </LabelText>
-          )}
-
           {contractData.instantiateInfo.ibcPortId && (
             <LabelText label="IBC Port ID">
               {contractData.instantiateInfo.ibcPortId}
@@ -112,10 +119,7 @@ export const InstantiateInfo = ({ contractData }: InstantiateInfoProps) => {
   };
 
   return (
-    <Flex direction="column" gap={6} w="180px">
-      <Heading as="h6" variant="h6">
-        Instantiate Info
-      </Heading>
+    <Flex direction="column" gap={6} w="250px">
       {renderDataFound()}
     </Flex>
   );